chore(idea): refactor plugin packages
