+++ packages/plugin-idea/src/main/kotlin/dev/elide/intellij/action/ElideRunAnythingProvider.kt
@@ -8,9 +8,9 @@ import com.intellij.ide.actions.runAnything.getPath
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.externalSystem.util.ExternalSystemApiUtil
 import dev.elide.intellij.Constants
-import dev.elide.intellij.execution.ElideExecutionService
-import dev.elide.intellij.project.data.elideProjectIndex
-import dev.elide.intellij.project.data.fullCommandLine
+import dev.elide.intellij.service.ElideExecutionService
+import dev.elide.intellij.service.elideProjectIndex
+import dev.elide.intellij.project.model.fullCommandLine
 import dev.elide.intellij.settings.ElideSettings
 import javax.swing.Icon
