+++ packages/plugin-idea/src/main/kotlin/dev/elide/intellij/cli/ElideCommandLineInfo.kt
@@ -9,8 +9,8 @@ import com.intellij.openapi.observable.util.createTextModificationTracker
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.ModificationTracker
 import dev.elide.intellij.Constants
-import dev.elide.intellij.project.data.elideProjectIndex
-import dev.elide.intellij.project.data.fullCommandLine
+import dev.elide.intellij.service.elideProjectIndex
+import dev.elide.intellij.project.model.fullCommandLine
 import javax.swing.Icon
 
 /** Extension used to provide completion suggestions and assistance for Elide run configurations. */
